requirements.txt
pandas
numpy
matplotlib
kagglehub
scikit-learn
python-graphviz